Moving some types around to tidy up the main files. Hopefully making it easier to grasp.
