+++ Dapper.Contrib/Attributes/ComputedAttribute.cs
@@ -1,0 +1,12 @@
+﻿using System;
+
+namespace Dapper.Contrib.Extensions
+{
+    /// <summary>
+    /// Specifies that this is a computed column.
+    /// </summary>
+    [AttributeUsage(AttributeTargets.Property)]
+    public class ComputedAttribute : Attribute
+    {
+    }
+}
