+++ Dapper.Contrib/Attributes/ExplicitKeyAttribute.cs
@@ -1,0 +1,12 @@
+﻿using System;
+
+namespace Dapper.Contrib.Extensions
+{
+    /// <summary>
+    /// Specifies that this field is a explicitly set primary key in the database
+    /// </summary>
+    [AttributeUsage(AttributeTargets.Property)]
+    public class ExplicitKeyAttribute : Attribute
+    {
+    }
+}
