+++ Dapper.Contrib/Attributes/KeyAttribute.cs
@@ -1,0 +1,12 @@
+﻿using System;
+
+namespace Dapper.Contrib.Extensions
+{
+    /// <summary>
+    /// Specifies that this field is a primary key in the database
+    /// </summary>
+    [AttributeUsage(AttributeTargets.Property)]
+    public class KeyAttribute : Attribute
+    {
+    }
+}
