+++ Dapper.Contrib/Attributes/TableAttribute.cs
@@ -1,0 +1,25 @@
+﻿using System;
+
+namespace Dapper.Contrib.Extensions
+{
+    /// <summary>
+    /// Defines the name of a table to use in Dapper.Contrib commands.
+    /// </summary>
+    [AttributeUsage(AttributeTargets.Class)]
+    public class TableAttribute : Attribute
+    {
+        /// <summary>
+        /// Creates a table mapping to a specific name for Dapper.Contrib commands
+        /// </summary>
+        /// <param name="tableName">The name of this table in the database.</param>
+        public TableAttribute(string tableName)
+        {
+            Name = tableName;
+        }
+
+        /// <summary>
+        /// The name of the table in the database
+        /// </summary>
+        public string Name { get; set; }
+    }
+}
