+++ Dapper.Contrib/Attributes/WriteAttribute.cs
@@ -1,0 +1,25 @@
+﻿using System;
+
+namespace Dapper.Contrib.Extensions
+{
+    /// <summary>
+    /// Specifies whether a field is writable in the database.
+    /// </summary>
+    [AttributeUsage(AttributeTargets.Property)]
+    public class WriteAttribute : Attribute
+    {
+        /// <summary>
+        /// Specifies whether a field is writable in the database.
+        /// </summary>
+        /// <param name="write">Whether a field is writable in the database.</param>
+        public WriteAttribute(bool write)
+        {
+            Write = write;
+        }
+
+        /// <summary>
+        /// Whether a field is writable in the database.
+        /// </summary>
+        public bool Write { get; }
+    }
+}
