+++ Dapper/TypeExtensions.cs
@@ -1,6 +1,6 @@
 ﻿using System;
-using System.Reflection;
 using System.Collections.Generic;
+using System.Reflection;
 
 namespace Dapper
 {
